feat: use vue charts instead of manual chart js for folder size history
